+++ packages/cesium/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @summeruse/ol
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies [dbeece1]
+- Updated dependencies [d3d0faf]
+  - @summeruse/common@0.1.0
+
 ## 0.1.0
 
 ### Minor Changes